% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-model.R
\name{add_model}
\alias{add_model}
\title{Add model elements to stan model}
\usage{
add_model(stan_model, string)
}
\arguments{
\item{stan_model}{A object of class 'stan_model'}

\item{string}{model element in text}
}
\description{
Introduce a model element to stan model
}
\examples{
stan_model <- create_stan_model()
add_model(stan_model, string = 'y ~ normal(0,1)')

}
