% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-model.R
\name{add_parameter}
\alias{add_parameter}
\title{Add parameter to stan model}
\usage{
add_parameter(stan_model, ...)
}
\arguments{
\item{stan_model}{A object of class 'stan_model'}

\item{...}{Mandatory arguments are declaration and name. minimum, maximum, and length are optional.}
}
\description{
Introduce a parameter declaration to stan model
}
\examples{
stan_model <- create_stan_model()
add_parameter(stan_model, declaration = "int", name = "test",
minimum = 3, maximum = 5, length = 500)

}
