% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfplot.R
\name{cdfplot}
\alias{cdfplot}
\alias{cdfplot.stanOutput}
\alias{cdfplot.stanPosterior}
\alias{cdfplot.stanPosteriors}
\title{Cumulative density plot function}
\usage{
cdfplot(x, ...)

\method{cdfplot}{stanPosterior}(object, figure_dir = "./")

\method{cdfplot}{stanPosteriors}(object)

\method{cdfplot}{stanOutput}(object, figure_dir = "./")
}
\description{
Plots posterior densities for \code{stanOutput} (all estimated parameters with chains), \code{stanPosterior} 
(all parameters in object) and \code{stanPosteriors} (all parameters with alternate models) objects.
}

