% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{create_data_block}
\alias{create_data_block}
\title{Create data block}
\usage{
create_data_block(stan_data)
}
\arguments{
\item{stan_data}{a list with stan data}
}
\description{
This function creates data block for stan models
from a stan data list
}
