% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatize-data-block.R
\name{create_model_from_data}
\alias{create_model_from_data}
\title{Automatize creation of data block}
\usage{
create_model_from_data(stan_data)
}
\arguments{
\item{stan_data}{a list with stan data}
}
\description{
This function automatizes the creation of  data block for stan models
from a stan data list
}
