% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-model.R
\name{create_stan_model}
\alias{create_stan_model}
\title{Create stan model}
\usage{
create_stan_model(data = list(), parameters = list(), model = list())
}
\arguments{
\item{data}{List of data declarations}

\item{parameters}{List of parameters}

\item{model}{Model definition}
}
\description{
This function returns a stan_model object.
}
