% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{densityplot}
\alias{densityplot}
\title{Density plot function}
\usage{
densityplot(x, ...)
}
\description{
Plots posterior densities for \code{stanOutput} (all estimated parameters with chains), \code{stanPosterior} 
(all parameters in object) and \code{stanPosteriors} (all parameters with alternate models) objects.
}

