% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-stanPosterior.R
\name{flatten}
\alias{flatten}
\alias{flatten.array}
\alias{flatten.matrix}
\alias{flatten.stanPosterior}
\title{Flatten stanPosterior object}
\usage{
flatten(object, ...)

\method{flatten}{stanPosterior}(object)

\method{flatten}{array}(object, par = character())

\method{flatten}{matrix}(object, par = character())
}
\description{
Converts \code{stanPosterior-class} object into a labelled data.frame suitable for plotting or printing
}

