% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histplot.R
\name{histplot}
\alias{histplot}
\title{Density plot function}
\usage{
histplot(x, ...)
}
\description{
Plots posterior densities for \code{stanOutput} (all estimated parameters with chains), \code{stanPosterior} 
(all parameters in object) and \code{stanPosteriors} (all parameters with alternate models) objects.
}

