% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\alias{posterior.character}
\alias{posterior.stanOutput}
\title{Extract permuted posterior distributions for selected model outputs}
\usage{
posterior(x, ...)

\method{posterior}{stanOutput}(object, pars = object@outputs)

\method{posterior}{character}(models, pars, path = rep(".", length(models)),
  ...)
}
\description{
These functions can be used to extract a list of output values from a \code{stanOutput-class} object in the current workspace.
If supplied with a character vector will search for saved \code{*.rds} objects and load them.
}

