% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-stanPosterior.R
\name{print.stanPosterior}
\alias{print.stanPosterior}
\title{Print functions for stanPosterior objects}
\usage{
\method{print}{stanPosterior}(object, write.to.word = FALSE)
}
\description{
Print functions for stanPosterior objects
}

