% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-stan-map.R
\name{read_stan_map}
\alias{read_stan_map}
\title{Read MAP csv files created by CmdStan or RStan}
\usage{
read_stan_map(csvfiles)
}
\arguments{
\item{csvfiles}{A character vector of paths to CSV files of MAP output}
}
\description{
Read MAP csv files created by CmdStan or RStan
}

