% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-restart.R
\name{restart}
\alias{restart}
\title{Extract permuted posterior sample from mcmc chain}
\usage{
restart(object, ...)
}
\arguments{
\item{object}{\code{stanOutput} class object containing MCMC samples}
}
\description{
This can be used for restarting another chain
}

