% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-stan.R
\name{run_stan}
\alias{run_stan}
\title{Run Stan model}
\usage{
run_stan(stan_model, stan_data, ...)
}
\arguments{
\item{stan_model}{stan_model object}

\item{stan_data}{data in stan template}

\item{...}{other arguments. Check out stan manual}
}
\description{
Run stan model with data and stan_model inputs
}
