% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-extract.R
\name{stan_extract}
\alias{stan_extract}
\title{Extract model fits and load into stanOutput object class}
\usage{
stan_extract(data = TRUE, map = FALSE, mcmc = FALSE,
  variational = FALSE, parameters = character(), outputs = character(),
  model = character(), mcmc_files = "all", path = ".")
}
\arguments{
\item{data}{read data into object?}

\item{map}{read map results into object?}

\item{mcmc}{read mcmc chains into object?}

\item{model}{a name or label for the model being extracted}

\item{map}{read variational results into object?}
}
\description{
Extract model fits and load into stanOutput object class
}

