% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanOutput-class.R
\docType{class}
\name{stanOutput-class}
\alias{stanOutput-class}
\title{stanOutput object class for storing outputs from a particular stan model run}
\description{
This is the central class for the \code{stanUtils} package, from which all other functionality is derived. It contains
the data and outputs from a particular model run, plus the name of the model run, a character vector of parameters and a 
character vector of model outputs.
}
\details{
Estimated \code{parameters} are stored as separate MCMC chains which can be overlaid during calls to \code{traceplot} etc. and are
useful for checking model convergence. Model \code{outputs} contain permuted MCMC outputs (all chains combined). These can be used for comparing
different model runs, but are primarily intended to provide useable model outputs. Estimated parameters should contain all estimated 
parameters in the \code{parameters {}} block of the stan model. The specification of odel ouputs is not restrictive and 
can contain any combination of variables specified anywhere in the model.

When initialising a \code{stanOutput} object, parameters and outputs must be specified as input character vectors. When using
\code{stan_extract} they can also be provided as a \code{*.par} file.
}

