% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanPosterior-class.R
\docType{class}
\name{stanPosterior-class}
\alias{stanPosterior-class}
\title{stanPosterior object class}
\description{
Contains MCMC samples from a subset of parameters extracted from an \code{stanOutput} object.
}

