% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R, R/histplot.R, R/traceplot.R
\name{densityplot.stanPosterior}
\alias{densityplot.stanOutput}
\alias{densityplot.stanPosterior}
\alias{densityplot.stanPosteriors}
\alias{histplot.stanOutput}
\alias{histplot.stanPosterior}
\alias{histplot.stanPosteriors}
\alias{traceplot}
\alias{traceplot.stanOutput}
\alias{traceplot.stanPosterior}
\alias{traceplot.stanPosteriors}
\title{Traceplot function}
\usage{
\method{densityplot}{stanPosterior}(object)

\method{densityplot}{stanPosteriors}(object)

\method{densityplot}{stanOutput}(object, figure_dir = "./")

\method{histplot}{stanPosterior}(object, bins = 20)

\method{histplot}{stanPosteriors}(object, bins = 20)

\method{histplot}{stanOutput}(object, pars = object@parameters, bins = 20)

traceplot(x, ...)

\method{traceplot}{stanPosterior}(object)

\method{traceplot}{stanPosteriors}(object)

\method{traceplot}{stanOutput}(object, pars = object@parameters)
}
\description{
Plots traces for \code{stanOutput} (all estimated parameters with chains), \code{stanPosterior} 
(all parameters in object) and \code{stanPosteriors} (all parameters with alternate models) objects.
}

