% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle-data.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Create stan data}
\usage{
create_stan_data(data, ...)
}
\arguments{
\item{data}{tibble or data.frame object}

\item{...}{index}
}
\description{
This function creates a data object 
to be used in Stan
}
\examples{
df <- data.frame(test = c(rep("dog", 5), rep("cat", 6)))
df <- create_index(df, k = test)
df <- create_stan_data(df, k)

}
