% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-model.R
\name{generate_stan_code}
\alias{generate_stan_code}
\title{Generate stan code}
\usage{
generate_stan_code(stan_model)
}
\arguments{
\item{stan_model}{A object of class `stan_model`}
}
\description{
This functions gets a stan_model object
and generate stan code from it.
}
